# Optimize comparison in `Cmp()`