We can do it with a single comparison. The speed up for the bad case is ~30%.
```
name                   old time/op    new time/op    delta
_Cmp/single/big-8        13.1ns ± 4%    13.1ns ± 2%     ~     (p=0.887 n=10+9)
_Cmp/single/uint256-8    5.40ns ± 3%    4.69ns ± 3%  -13.17%  (p=0.000 n=10+9)

name                   old alloc/op   new alloc/op   delta
_Cmp/single/big-8         0.00B          0.00B          ~     (all equal)
_Cmp/single/uint256-8     0.00B          0.00B          ~     (all equal)

name                   old allocs/op  new allocs/op  delta
_Cmp/single/big-8          0.00           0.00          ~     (all equal)
_Cmp/single/uint256-8      0.00           0.00          ~     (all equal)
```