@fyrchik 2021-07-05
I am not sure if this is the right way to change benchmark (we get duration of 2 operations), but it seems simpler than making 2 more benchmarks.